Unuse fsnotify on wasm
